% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_cleaning_functions.R
\name{getMinMax}
\alias{getMinMax}
\title{Minimum and Maximum Values}
\usage{
getMinMax(
  dataframe,
  base_variables = c("PatientID", "StudyDate", "StudyNotes", "InterpreterID"),
  temp_addtl_variables = "all",
  numberTopBottom
)
}
\arguments{
\item{dataframe}{The dataframe to perform the function on (make sure desired
filters are applied) and that variables you want top/bottom values for are
numeric variables}

\item{base_variables}{The variables that identify which row each top/bottom
value comes from. These variables will not have top/bottom values}

\item{temp_addtl_variables}{These are the variables to find top/bottom values
for. If "all" is provided, the function will attempt to get all variables
that are not base_variables.}

\item{numberTopBottom}{This is the number of top/bottom values to get. That
is, if you want the top 10/ bottom 10 values, then numberTopBottom = 10}
}
\value{
A nested dataframe
}
\description{
A function that gets the top and bottom n values for each
  specified variable.
}
\details{
This should return a nested dataframe with a column of dataframes of
  top and bottom values for each variable.
}
